Move tips membership benchmark logic to its own module
